+++ tests/tests.py
@@ -48,8 +48,10 @@ localeCode = "en"
 # src folder during unit testing outside of the container).
 StatConfig._locale_directory = "src" + StatConfig._locale_directory
 
+fakedOptionalReposContributedToQueryResult = {'data': {'user': {'repositoriesContributedTo': {'totalCount': 3}}}}
+
 executedQueryResultsOriginal = [
-    {'data': {'user': {'followers': {'totalCount': 9}, 'following': {'totalCount': 7}, 'issues': {'totalCount': 81}, 'login': 'someuser', 'name': 'Firstname M. Lastname', 'pullRequests': {'totalCount': 289}, 'repositoriesContributedTo': {'totalCount': 3}, 'sponsorshipsAsMaintainer': {'totalCount': 7}, 'sponsorshipsAsSponsor': {'totalCount': 5}}}},
+    {'data': {'user': {'followers': {'totalCount': 9}, 'following': {'totalCount': 7}, 'issues': {'totalCount': 81}, 'login': 'someuser', 'name': 'Firstname M. Lastname', 'createdAt':'2011-05-01T15:46:30Z', 'pullRequests': {'totalCount': 289}, 'sponsorshipsAsMaintainer': {'totalCount': 7}, 'sponsorshipsAsSponsor': {'totalCount': 5}}}},
     
     {'data': {'user': {'contributionsCollection': {'totalCommitContributions': 3602, 'totalIssueContributions': 79, 'totalPullRequestContributions': 289, 'totalPullRequestReviewContributions': 315, 'totalRepositoryContributions': 18, 'restrictedContributionsCount': 105, 'contributionYears': [2021, 2020, 2019, 2018, 2017, 2016, 2015, 2014, 2013, 2012, 2011]}}}},
 
@@ -61,7 +63,7 @@ executedQueryResultsOriginal = [
     ]
 
 executedQueryResultsMultiPage = [
-    {'data': {'user': {'followers': {'totalCount': 9}, 'following': {'totalCount': 7}, 'issues': {'totalCount': 81}, 'login': 'someuser', 'name': 'Firstname M. Lastname', 'pullRequests': {'totalCount': 289}, 'repositoriesContributedTo': {'totalCount': 3}, 'sponsorshipsAsMaintainer': {'totalCount': 7}, 'sponsorshipsAsSponsor': {'totalCount': 5}}}},
+    {'data': {'user': {'followers': {'totalCount': 9}, 'following': {'totalCount': 7}, 'issues': {'totalCount': 81}, 'login': 'someuser', 'name': 'Firstname M. Lastname', 'createdAt':'2011-05-01T15:46:30Z', 'pullRequests': {'totalCount': 289}, 'sponsorshipsAsMaintainer': {'totalCount': 7}, 'sponsorshipsAsSponsor': {'totalCount': 5}}}},
     
     {'data': {'user': {'contributionsCollection': {'totalCommitContributions': 3602, 'totalIssueContributions': 79, 'totalPullRequestContributions': 289, 'totalPullRequestReviewContributions': 315, 'totalRepositoryContributions': 18, 'restrictedContributionsCount': 105, 'contributionYears': [2021, 2020, 2019, 2018, 2017, 2016, 2015, 2014, 2013, 2012, 2011]}}}},
 
@@ -86,7 +88,8 @@ class TestSomething(unittest.TestCase) :
                     executedQueryResults[0],
                     executedQueryResults[1],
                     executedQueryResults[2],
-                    executedQueryResults[4]
+                    executedQueryResults[4],
+                    contribToData = fakedOptionalReposContributedToQueryResult
                     )
                 #self.parsePriorYearStats(executedQueryResults[3])
         stats = NoQueries(True, False, 1000, set(), None)
@@ -104,7 +107,8 @@ class TestSomething(unittest.TestCase) :
                     executedQueryResults[0],
                     executedQueryResults[1],
                     executedQueryResults[2],
-                    executedQueryResults[4]
+                    executedQueryResults[4],
+                    contribToData = fakedOptionalReposContributedToQueryResult
                     )
                 #self.parsePriorYearStats(executedQueryResults[3])
         stats = NoQueries(True, False, 1000, set(), None)
@@ -122,7 +126,8 @@ class TestSomething(unittest.TestCase) :
                     executedQueryResults[0],
                     executedQueryResults[1],
                     executedQueryResults[2],
-                    executedQueryResults[4]
+                    executedQueryResults[4],
+                    contribToData = fakedOptionalReposContributedToQueryResult
                     )
                 #self.parsePriorYearStats(executedQueryResults[3])
         stats = NoQueriesMultipage(True, False, 1000, {"repo29", "repoDoesntExist"}, None)
@@ -140,7 +145,8 @@ class TestSomething(unittest.TestCase) :
                     executedQueryResults[0],
                     executedQueryResults[1],
                     executedQueryResults[2],
-                    executedQueryResults[4]
+                    executedQueryResults[4],
+                    contribToData = fakedOptionalReposContributedToQueryResult
                     )
                 #self.parsePriorYearStats(executedQueryResults[3])
         stats = NoQueriesMultipage(True, False, 1000, {"repo29", "repoDoesntExist"}, None)
@@ -160,7 +166,8 @@ class TestSomething(unittest.TestCase) :
                     executedQueryResults[0],
                     executedQueryResults[1],
                     executedQueryResults[2],
-                    executedQueryResults[4]
+                    executedQueryResults[4],
+                    contribToData = fakedOptionalReposContributedToQueryResult
                     )
                 #self.parsePriorYearStats(executedQueryResults[3])
         stats = NoQueries(True, False, 1000, set(), None)
@@ -330,7 +337,8 @@ class TestSomething(unittest.TestCase) :
                     executedQueryResults[0],
                     executedQueryResults[1],
                     executedQueryResults[2],
-                    executedQueryResults[4]
+                    executedQueryResults[4],
+                    contribToData = fakedOptionalReposContributedToQueryResult
                     )
                 #self.parsePriorYearStats(executedQueryResults[3])
         stats = NoQueries(True, False, 100, set(), "FavoriteRepo")