+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2026-07-17
+## [Unreleased] - 2026-07-20
 
 ### Added
 
@@ -15,6 +15,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 
 ### Fixed
+* Fetch total number of commits from REST API
+* Split repositoriesContributedTo as separate query and treat as optional (i.e., just drop row if it fails)
+* Additional optimizations to GraphQL queries
 
 ### Dependencies